+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Updated package-lock.json to fix a potential vulnerability due to the merge dependency
+- Add ecdhCurve: 'auto' in the agent options to fix a bug in Node 8 (https://github.com/nodejs/node/issues/19359)
 
 ## [0.2.3] - 2018-10-29
 ### Changed